Move ban classes to ban package
