+++ src/main/java/org/spongepowered/api/service/ban/BanType.java
@@ -22,31 +22,22 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.util.ban;
+package org.spongepowered.api.service.ban;
 
-import org.spongepowered.api.Sponge;
+import org.spongepowered.api.CatalogType;
+import org.spongepowered.api.util.annotation.CatalogedBy;
 
-import java.util.function.Supplier;
-
-public class BanTypes {
-
-    // SORTFIELDS:ON
-
-    /**
-     * Represents a {@link Ban.Ip}.
-     */
-    public static final Supplier<BanType> IP = Sponge.getRegistry().getCatalogRegistry().provideSupplier(BanType.class, "ip");
+/**
+ * Represents the possible types of bans.
+ */
+@CatalogedBy(BanTypes.class)
+public interface BanType extends CatalogType {
 
     /**
-     * Represents a {@link Ban.Profile}.
+     * Gets the {@link Ban} class that this type represents.
+     *
+     * @return The ban class
      */
-    public static final Supplier<BanType> PROFILE = Sponge.getRegistry().getCatalogRegistry().provideSupplier(BanType.class, "profile");
-
-    // SORTFIELDS:OFF
-
-    // Suppress default constructor to ensure non-instantiability.
-    private BanTypes() {
-        throw new AssertionError("You should not be attempting to instantiate this class.");
-    }
+    Class<? extends Ban> getBanClass();
 
 }
